3D converter
IVCon is an executable C++ program that can read and write a variety of
3D graphics file formats, converting from one to the other.

This is extended version of John Burkardt's great, but buggy tool. It
Supports many 3D formats, and a lot of 3D scene elements.

Supported formats include 3DS, ASE, BYU, DXF, GMOD, HRC, IV, OBJ, OFF,
POV, SMF, STL/STLA/STLB, TEC, TRI/TRIA/TRIB, UCD, VLA, WRL/VRML, XGL and
more.

The program supports many elements of 3D scene, but specific converters
may not support all of them. There are two possible reasons of that: one
is that the support is not fully programmed inside this tool, second is
that the format may not have possibilities to support the item. Good
example is the ASC format - which is an ASCII cloud of vertices, and
supports only bare points in 3D space.

The program is free and open source.